Move protectForm to lib/protect-form
