+++ client/me/profile/index.jsx
@@ -9,7 +9,7 @@ import debugFactory from 'debug';
  * Internal dependencies
  */
 import MeSidebarNavigation from 'me/sidebar-navigation';
-import { protectForm } from 'components/hoc/protect-form';
+import { protectForm } from 'lib/protect-form';
 import formBase from 'me/form-base';
 import FormButton from 'components/forms/form-button';
 import FormFieldset from 'components/forms/form-fieldset';
