+++ client/my-sites/site-settings/form-jetpack-monitor.jsx
@@ -16,7 +16,7 @@ var config = require( 'config' ),
 	dirtyLinkedState = require( 'lib/mixins/dirty-linked-state' ),
 	SectionHeader = require( 'components/section-header' ),
 	Button = require( 'components/button' );
-import { protectForm } from 'components/hoc/protect-form';
+import { protectForm } from 'lib/protect-form';
 
 module.exports = protectForm( React.createClass( {
